[Refactor] Refactor namespaces and update project structure
This commit refactors the namespaces from `SmartMenuOptim.API.Data` to `SmartMenuOptim.Shared.Data`, promoting a shared data model. The `AppDbContext` and `DbSeeder` classes have been moved to the new namespace, and relevant using directives have been updated across multiple files, including controllers and Razor components.

Additionally, the project file has been modified to include package references for `Microsoft.EntityFrameworkCore` and `Newtonsoft.Json`, enhancing functionality. These changes aim to improve code organization and facilitate better data handling across the application.

[Refactor] Refactor DTO folder structure in project file

Updated `SmartMenuOptim.Shared.csproj` to remove the
`Data\Entities\Dtos\` folder and include `Data\Dtos\`
for improved organization of Data Transfer Objects.

[Refactor]  Refactor project structure for model classes

Moved model classes from `SmartMenuOptim.Shared.Models`
to `SmartMenuOptim.Shared.Data.Entities` namespace.
Updated all references in controllers, services, and data
models accordingly. Modified project file to reflect
new folder structure and added necessary using directives
for the new namespace.
