+++ SmartMenuOptim.API/Controllers/AiController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using SmartMenuOptim.Shared.Models;
+using SmartMenuOptim.Shared.Data.Entities;
+using SmartMenuOptim.Shared.Data.Entities;
 
 namespace SmartMenuOptim.API.Controllers
 {
