+++ SmartMenuOptim.API/Data/DbSeeder.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json.Linq;
-using SmartMenuOptim.Shared.Models;
+using SmartMenuOptim.Shared.Data.Entities;
 
-namespace SmartMenuOptim.API.Data
+namespace SmartMenuOptim.Shared.Data
 {
     public static class DbSeeder
     {
