+++ SmartMenuOptim.Server/Components/Pages/Dashboard.razor
@@ -1,5 +1,6 @@
 ﻿@page "/dashboard"
 @using SmartMenuOptim.Shared.Data
+@using SmartMenuOptim.Shared.Data.Entities
 @using SmartMenuOptim.Shared.Models
 @inject NavigationManager NavigationManager
 @inject ISaleRecordService SaleRecordService
