+++ SmartMenuOptim.Server/Components/Pages/Insights.razor
@@ -1,6 +1,6 @@
 ﻿@page "/insights"
 @using SmartMenuOptim.Server.Services
-@using SmartMenuOptim.Shared.Models
+@using SmartMenuOptim.Shared.Data.Entities
 @inject IAIService AIService // Injecting the AI service to get recommendations
 @inject IReviewService ReviewService // Injecting the review service to get reviews
 @inject ISaleRecordService SaleRecordService // Injecting the sale record service to get sales data
