+++ SmartMenuOptim.Server/Components/Pages/Reviews.razor
@@ -1,6 +1,6 @@
 ﻿@page "/reviews"
 @using SmartMenuOptim.Server.Services
-@using SmartMenuOptim.Shared.Models
+@using SmartMenuOptim.Shared.Data.Entities
 @inject IAIService AIService
 @inject IReviewService ReviewService
