+++ SmartMenuOptim.Server/Services/IReviewService.cs
@@ -1,4 +1,4 @@
-﻿using SmartMenuOptim.Shared.Models;
+﻿using SmartMenuOptim.Shared.Data.Entities;
 
 internal interface IReviewService
 {
