+++ SmartMenuOptim.Server/Services/ReviewService.cs
@@ -1,4 +1,4 @@
-﻿using SmartMenuOptim.Shared.Models;
+﻿using SmartMenuOptim.Shared.Data.Entities;
 
 internal class ReviewService : IReviewService
 {
